+++ oggm/utils/_funcs.py
@@ -780,3 +780,82 @@ def shape_factor_adhikari(widths, heights, is_rectangular):
     shape_factors[np.isnan(shape_factors)] = 1.
 
     return shape_factors
+
+
+def cook_rgidf(gi_gdf, o1_region, o2_region='01', version='60', ids=None,
+               bgndate='20009999', id_suffix='', assign_column_values=None):
+    """Convert a glacier inventory into a dataset looking like the RGI (OGGM ready).
+
+    Parameters
+    ----------
+    gi_gdf : :py:geopandas.GeoDataFrame
+        the GeoDataFrame of the user's glacier inventory.
+    o1_region : str
+        Glacier RGI region code, which is important for some OGGM applications.
+        For example, oggm.shop.its_live() need it to locate the right dataset.
+        Needs to be assigned.
+    o2_region :  str or list of str
+        Glacier RGI subregion code (default: 01)
+    bgndate : str or list of str
+        The date of the outlines. This is quite important for the glacier
+        evolution runs, which start at the RGI date. Format: ``'YYYYMMDD'``,
+        (MMDD is not used).
+    version : str
+        Glacier inventory version code, which is necessary to generate the RGIId.
+        The default is '60'.
+    ids : list of IDs as integers. The default is None, which generates the
+        RGI ids automatically following the glacier order.
+    id_suffix : str or None
+         Add a suffix to the glacier ids. The default is None, no suffix
+    assign_column_values : dict or None
+        Assign predefined values from the original data to the RGI dataframe.
+        dict format :
+            - key: name of the column in the original dataframe
+            - value: name of the column in the RGI-like file to assign the values to
+
+    Returns
+    -------
+    cooked_rgidf : :py:geopandas.GeoDataFrame
+        glacier inventory into a dataset looking like the RGI (OGGM ready)
+
+    """
+
+    # Check input
+    if ids is None:
+        ids = range(1, len(gi_gdf)+1)
+
+    # Construct a fake rgiid list following RGI format
+    str_ids = ['RGI{}-{}.{:05d}{}'.format(version, o1_region, int(i), id_suffix)
+               for i in ids]
+
+    # Check the coordination system.
+    # RGI use the geographic coordinate.
+    # So if the original glacier inventory is not in geographic coordinate,
+    # we need to convert both of the central point and the glacier outline
+    # to geographic coordinate (WGS 84)
+    if gi_gdf.crs != 'epsg:4326':
+        gi_gdf = gi_gdf.to_crs('epsg:4326')
+
+    # Calculate the central point of the glaciers
+    geom = gi_gdf.geometry.values
+    clon = gi_gdf.geometry.centroid.x
+    clat = gi_gdf.geometry.centroid.y
+
+    # Prepare data for the output GeoDataFrame
+    data = {'RGIId': str_ids, 'CenLon': clon, 'CenLat': clat, 'GLIMSId': '',
+            'BgnDate': bgndate, 'EndDate': '9999999',
+            'O1Region': o1_region, 'O2Region': o2_region,
+            'Area': -9999., 'Zmin': -9999., 'Zmax': -9999., 'Zmed': -9999.,
+            'Slope': -9999., 'Aspect': -9999, 'Lmax': -9999,
+            'Status': 0, 'Connect': 0, 'Form': 0, 'TermType': 0, 'Surging': 0,
+            'Linkages': 1, 'check_geom': None, 'Name': ''}
+
+    # Construct the output GeoDataFrame
+    cooked_rgidf = gpd.GeoDataFrame(data=data, geometry=geom, crs='epsg:4326')
+
+    # If there are specific column in the original glacier inventory we want to keep
+    if assign_column_values is not None:
+        for key, val in assign_column_values.items():
+            cooked_rgidf[val] = gi_gdf[key].values
+
+    return cooked_rgidf