+++ docs/whats-new.rst
@@ -13,6 +13,10 @@ Breaking changes
 Enhancements
 ~~~~~~~~~~~~
 
+- Added ``cook_rgidf()`` function in ``oggm.utils`` to simplify the use
+  of a non-RGI glacier inventory in OGGM (:pull:`1251`).
+  By `Li Fei <https://github.com/Keeptg>`_
+
 Bug fixes
 ~~~~~~~~~